HTML Tags are no longer so contrasting in the light view. However, comments return to having the high contrast background.
Comments are great for providing visual breaks to long stretches of code, and so get the same look at Markdown headings. This also fits with the aesthetic of a circuit board, where comments mark out what components are, or give some reminder text to the user of the device (pocket operator).

HTML tags should not have been so high contrast, and as I work more with templates, they should have been treated equally to Liquid or similar substitution punctuation (like f-strings in python). Liquid, meta.embedded, and meta.interpolated are similarly styled with a background highlight.